lss: implement LED logic
We would like a simple way to give feedback when LSS
is involved. Particularly, we would like to identify
a node that is not configured (factory settings), or
that is currently being configured.
Implement a special logic, which can be used to override
the standard LED behavior if your device does not have
a dedicated LED. After all, LSS can be seen as a
maintenance mode so it can be acceptable to deviate
from the standard.