remove the need to specify the number OD entries
The number of entries in the OD was specified in CO_OD.h. This is fine as long
as the CO_OD.h and CO_OD.c files are generated automatically, but is very error
prone in cases where these files are maintained manually.

Here we remove the need to specify that count up front, while still staying
compatible with existing code.